+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [1.1.0 - 2026-02-05]
+## [1.1.0 - 2026-02-06]
 ### Added
 - Dockerfile
 - Single Spa
@@ -132,3 +132,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Removed
 - Application Form
+
+### Fixed
+- Update the dialog stepper only if necessary